refactor(services): reorganize User services into specific subpackages
Move User related services into subpackage services.user.

Update imports in controllers, tests.
